Removed old Login UI
